+++ kodilla-patterns2/src/main/java/com/kodilla/patterns2/adapter/bookclassifier/librarya/Classifier.java
@@ -3,5 +3,5 @@ package com.kodilla.patterns2.adapter.bookclassifier.librarya;
 import java.util.Set;
 
 public interface Classifier {
-    int publicationYearMedian(Set<Book> bookSet);
+    double publicationYearMedian(Set<Book> bookSet);
 }
